feat: add rewards controller mock and fix tests

+++ src/interfaces/IRewardsController.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.8.18;
 
 interface IRewardsController {
-    function tierOf(uint256 tokenId) external view returns (uint256);
+    function tierOf(uint256 tokenId) external view returns (uint8);
 }
